% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processIndicator.R
\name{process_ANC}
\alias{process_ANC}
\title{Process women's dataset for ANC 4+ visits}
\usage{
process_ANC(wm)
}
\arguments{
\item{wm}{A women's data.frame (wm.sav) from MICS.}
}
\value{
A data.frame with columns:
\code{cluster}, \code{householdID}, \code{weight}, \code{strata},
\code{value}, \code{indicator}
}
\description{
Process women's dataset for ANC 4+ visits
}
