% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_geo_mics.R
\name{process_geo_mics}
\alias{process_geo_mics}
\title{Process geo for supported MICS country-year combinations}
\usage{
process_geo_mics(country.name, gps_path, out_path = NULL, year)
}
\arguments{
\item{country.name}{Character. Must match a supported handler key (aliases later).}

\item{gps_path}{Path to GPS .shp}

\item{out_path}{Optional .RData output path}

\item{year}{Numeric/integer or character. Used to select country-year handler.}
}
\value{
A list with cluster.info, admin.info1, admin.info2, admin0/admin1/admin2, geo
}
\description{
Users provide GPS shapefile; this constructs cluster/admin objects
needed by surveyPrev and SUMMER.
}
\details{
Supported:
\itemize{
\item Honduras 2019
\item Nigeria 2016
\item Nigeria 2021
\item Ghana 2017-18 (use year = 2017 or 2018)
\item Lao PDR 2017
\item Gambia 2018
\item Sierra Leone 2017
\item Madagascar 2018
\item Zimbabwe 2019
\item Thailand 2022
\item Malawi 2019-20 (use year = 2019 or 2020)
}
}
