% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processIndicator.R
\name{process_NMR}
\alias{process_NMR}
\title{Process birth history dataset for NMR}
\usage{
process_NMR(bh)
}
\arguments{
\item{bh}{A birth history data.frame (bh.sav) from MICS.}
}
\value{
A data.frame with columns:
\code{cluster}, \code{householdID}, \code{weight}, \code{strata},
\code{value}, \code{indicator}
}
\description{
Process birth history dataset for NMR
}
